+++ spec/models/services/tumblr_spec.rb
@@ -22,7 +22,7 @@ describe Services::Tumblr, type: :model do
 
     before do
       user.services << service
-      stub_request(:get, "http://api.tumblr.com/v2/user/info").to_return(status: 200, body: user_info)
+      stub_request(:get, "https://api.tumblr.com/v2/user/info").to_return(status: 200, body: user_info)
     end
 
     context "with multiple blogs" do
@@ -32,8 +32,8 @@ describe Services::Tumblr, type: :model do
                     {
                       user: {
                         blogs: [
-                          {primary: false, url: "http://foo.tumblr.com"},
-                          {primary: true, url: "http://bar.tumblr.com"}
+                          {primary: false, url: "https://foo.tumblr.com"},
+                          {primary: true, url: "https://bar.tumblr.com"}
                         ]
                       }
                     }
@@ -41,7 +41,7 @@ describe Services::Tumblr, type: :model do
       end
 
       it "posts a status message to the primary blog and stores the id" do
-        stub = stub_request(:post, "http://api.tumblr.com/v2/blog/bar.tumblr.com/post")
+        stub = stub_request(:post, "https://api.tumblr.com/v2/blog/bar.tumblr.com/post")
                .with(post_request).to_return(post_response)
 
         expect(post).to receive(:tumblr_ids=).with({"bar.tumblr.com" => post_id}.to_json)
@@ -53,10 +53,10 @@ describe Services::Tumblr, type: :model do
     end
 
     context "with a single blog" do
-      let(:user_info) { {response: {user: {blogs: [{url: "http://foo.tumblr.com"}]}}}.to_json }
+      let(:user_info) { {response: {user: {blogs: [{url: "https://foo.tumblr.com"}]}}}.to_json }
 
       it "posts a status message to the returned blog" do
-        stub = stub_request(:post, "http://api.tumblr.com/v2/blog/foo.tumblr.com/post")
+        stub = stub_request(:post, "https://api.tumblr.com/v2/blog/foo.tumblr.com/post")
                .with(post_request).to_return(post_response)
 
         service.post(post)
@@ -80,7 +80,7 @@ describe Services::Tumblr, type: :model do
   describe "#delete_from_service" do
     it "removes posts from tumblr" do
       tumblr_ids = {"foodbar.tumblr.com" => post_id}.to_json
-      stub = stub_request(:post, "http://api.tumblr.com/v2/blog/foodbar.tumblr.com/post/delete")
+      stub = stub_request(:post, "https://api.tumblr.com/v2/blog/foodbar.tumblr.com/post/delete")
              .with(body: {"id" => post_id}).to_return(status: 200)
 
       service.delete_from_service(tumblr_ids: tumblr_ids)